Clean up provided code
The provided code includes the entire html document and my template is designed
to only place content in the <body> tag. Merging the two is all that is being
done, for now, to get the HTML document into a workable version.

Modify a few other items, such as removing default styles I don't like, and
inserting a comment, to a comment, in the original HTML. Moved the media
query to the end of the style block and replaced the url for the hamburger menu
with the one supplied in the comments.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>