Add javascript for menu controls
Used one method, adding and removing a class, for controlling the mobile menu.

Used a second method, changing the display property between block and none to
control the second drop-down menu. On the Solutions entry. The process is
written to control all three drop-downs, with the first and third commented out.

Used a third method, adding and removing the DOM fragment from the document's
DOM tree, to control the first menu. Again, written to control all three menus,
with the lines for the second and third menus commented out. This version is
also coded in an OOP style using a constructor and methods.

Used the CSS-based :hover selector to control the third menu. With something as
simple as these menus I prefer a CSS approach, reserving JavaScript for complex
and intricate tasks. The selector is strange, since it is limited to one of a
group of three which could (should?) be so controlled. Comments by the selector
give the normal selector to use if it were controlling all three menus.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>