Cosmetic changes
Added <span> tags around the buttons for the drop-down menus. I didn't like the
link pointer showing up in all the white-space associated with the menus. Only
want that to show over, or very near, the text of the links, or inside buttons.

Changing the cursor sensitivity suggested also adjusting the element to attach
the event trigger to as well. Changed in both versions of the JS actions and in
the CSS :hover selector.

Added the correct font-size to the 'title' of the drop-down menus.

Added text to the blank main section of the page to "discuss" the DOM placement
of drop-down menus and to report the issue, or bug, with the provided solution
to the homework.

Marked as submitted.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>